Dockerise Dave.
* Environment Variables!

Made the launcher/core work with environment variables, since we're going to make the docker container use an env.list. The reddit code still stands to be modified.

* Reddit client now uses environvars.

Also added usage to exit message on none-provided vars. Leaving the else in just in case but made a modification in it.

* Add dockerfile.

And make sure env.list isn't commited so that secrets are protected. It works! Just need to update documentation.

* Added .dockerignore

* Mod to dockerignore

Ignore readme.md

* Updated readme to reflect changes.

Included instructions for docker and some modification to the rest of the docs.