Removed !bothelp and fixed an error.
Turns out !help works and uses docstrings. Launching with one arg was calling with the wrong thing. It now doesn't, but still doesn't quite work as expected. I'm not personally launching it like that so I'll fix it next release, but I don't thing the fix will make it into this release.